requirements.txt
yfinance
pandas
psycopg2-binary
python-dotenv
sqlalchemy
streamlit
plotly
